sr(api): move user-recommended-training-repository in devcomp
