+++ api/lib/infrastructure/repositories/index.js
@@ -5,7 +5,7 @@ import * as trainingRepository from '../../../src/devcomp/infrastructure/reposit
 import * as trainingTriggerRepository from '../../../src/devcomp/infrastructure/repositories/training-trigger-repository.js';
 import * as tutorialEvaluationRepository from './tutorial-evaluation-repository.js';
 import * as tutorialRepository from './tutorial-repository.js';
-import * as userRecommendedTrainingRepository from './user-recommended-training-repository.js';
+import * as userRecommendedTrainingRepository from '../../../src/devcomp/infrastructure/repositories/user-recommended-training-repository.js';
 
 import { fromDatasourceObject } from '../adapters/solution-adapter.js';
 import { getCorrection } from '../../domain/services/solution-service-qrocm-dep.js';
