+++ api/src/devcomp/infrastructure/repositories/user-recommended-training-repository.js
@@ -1,6 +1,6 @@
-import { knex } from '../../../db/knex-database-connection.js';
-import { DomainTransaction } from '../DomainTransaction.js';
-import { UserRecommendedTraining } from '../../domain/read-models/UserRecommendedTraining.js';
+import { knex } from '../../../../db/knex-database-connection.js';
+import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { UserRecommendedTraining } from '../../../../lib/domain/read-models/UserRecommendedTraining.js';
 
 const TABLE_NAME = 'user-recommended-trainings';
