+++ api/tests/devcomp/integration/infrastructure/repositories/user-recommended-training-repository_test.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
-import { databaseBuilder, expect, knex } from '../../../test-helper.js';
-import * as userRecommendedTrainingRepository from '../../../../lib/infrastructure/repositories/user-recommended-training-repository.js';
-import { UserRecommendedTraining } from '../../../../lib/domain/read-models/UserRecommendedTraining.js';
+import { databaseBuilder, expect, knex } from '../../../../test-helper.js';
+import * as userRecommendedTrainingRepository from '../../../../../src/devcomp/infrastructure/repositories/user-recommended-training-repository.js';
+import { UserRecommendedTraining } from '../../../../../lib/domain/read-models/UserRecommendedTraining.js';
 
 describe('Integration | Repository | user-recommended-training-repository', function () {
   describe('#save', function () {
